Creo que esto soluciona el problema que mando Robert

+++ booktube/src/main/java/com/booktube/model/Comment.java
@@ -34,7 +34,7 @@ public class Comment implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "COMMENT_ID")
-	private Integer id;
+	private Long id;
 	
 	@ManyToOne(cascade=CascadeType.ALL, fetch = FetchType.LAZY)
 	@JoinColumn(name="USER_ID", updatable = false)
@@ -58,7 +58,7 @@ public class Comment implements Serializable {
 		
 	}
 	
-	public Comment(Integer id, User user, Book book, String text) {
+	public Comment(Long id, User user, Book book, String text) {
 		this.id = id;
 		this.user = user;
 		this.book = book;
@@ -82,11 +82,11 @@ public class Comment implements Serializable {
 		this.book = book;
 	}
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
